+++ Changelog.md
@@ -13,6 +13,7 @@ OpenCore Changelog
 - Added kext blocker `Exclude` strategy for mkext
 - Re-enabled AudioDxe failover to protocol GET mode for systems such as Acer E5 where it works when DisconnectHda doesn't
 - Added `FirmwareSettingsEntry.efi` driver which adds menu entry to reboot into UEFI firmware settings
+- Enabled use of picker shortcut keys which are read out in OpenCanopy when using `PickerAudioAssist`
 
 #### v0.9.7
 - Updated recovery_urls.txt